Fixed the weird behavior of slider dots and buttons
When we clicked on dot no. 3 and then we click right it spossed to go slide no. 1 but it goes to no. 2.
I just added current slide settings to the dot function.
```js
  dotContainer.addEventListener('click', e => {
    if (e.target.classList.contains('dots__dot')) {
      currentSlide = e.target.dataset.slide;
      goToSlides(e.target.dataset.slide);
    }
  });
```

And now you no need to active function again and  again just called one time in `goToSlides` function. Active class is set by the help of `currentSlide` variable